AppError and ValidationError moved to dtos.error package
