+++ src/main/java/org/blbulyandavbulyan/blog/controllers/handling/ValidationControllerAdvice.java
@@ -5,7 +5,7 @@ import jakarta.validation.ConstraintViolationException;
 import jakarta.validation.ElementKind;
 import jakarta.validation.Path;
 import lombok.extern.slf4j.Slf4j;
-import org.blbulyandavbulyan.blog.exceptions.ValidationError;
+import org.blbulyandavbulyan.blog.dtos.error.ValidationError;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
